requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
sentence-transformers
streamlit